backend/requirements.txt
flask
flask-cors
pypdf2
spacy
sentence-transformers
torch
google-generativeai
python-dotenv
